Refactor shared error message generation -> validationUtils.js
